backend/requirements.txt
fastapi
mysql-connector-python
cors
hypercorn
